Tomorrow day flight checking
In the previous version, the algorithm won't check the flight takeoff
in the next day. Which results in the following problem, if someone
arrives at some airport where as no flight takesoff after his arrival in
that day, but some flight takes off in the following day and it's a
favourable choice, this node will be abondaned by the algorithm. The
current version evaluate all possible choices of all today's flight and
tomorrow's, but it'll slow down the searching process. I guess we might
try some way to accelerate this process.