Complete airport clawer
Complete a simplified version of airport clawer, which is able to claw
Municipality, Prefecture and it's children airports' ICAL and IATA
abbrivation. There are some incompleteness. First we discard all datas
that contains character that exceeds Uncode 8 standard, for I didn't
find city names like that in mileage database. Second we currently don't
have the option to assure unique contraint or to alter a data in the
database.
We may going to clear the entire table and re-claw the website to update
the database, since this table won't need a frequent update.
entire table